Include the HTML body in the JSON message data
